+++ flang/include/flang/Optimizer/OpenMP/Passes.td
@@ -73,6 +73,28 @@ def GlobalFilteringPass : Pass<"omp-global-filtering"> {
   ];
 }
 
+def HostOpFilteringPass : Pass<"omp-host-op-filtering"> {
+  let summary =
+      "Removes host-only operations from host functions retained by the "
+      "function filtering pass when compiling for a target device.";
+  let description = [{
+    Removes operations from host functions that are not placed inside of an
+    `omp.target` region when compiling for the target device.
+    
+    It must run after any passes that might introduce new host operations,
+    especially if they belong to the 'omp' dialect, like `LowerWorkdistribute`.
+    This ensures all potentially problematic host operations are accounted for
+    prior to LLVM IR translation.
+
+    This pass works at the FIR level of abstraction.
+  }];
+  let dependentDialects = [
+    "mlir::arith::ArithDialect",
+    "mlir::func::FuncDialect",
+    "fir::FIROpsDialect"
+  ];
+}
+
 def DoConcurrentConversionPass : Pass<"omp-do-concurrent-conversion", "mlir::ModuleOp"> {
   let summary = "Map `DO CONCURRENT` loops to OpenMP worksharing loops.";
