+++ flang/test/Fir/basic-program.fir
@@ -85,6 +85,7 @@ func.func @_QQmain() {
 // PASSES-NEXT:   ConvertHLFIRtoFIR
 // PASSES-NEXT:   LowerWorkshare
 // PASSES-NEXT:   LowerWorkdistribute
+// PASSES-NEXT:   HostOpFilteringPass
 // PASSES-NEXT:   CSE
 // PASSES-NEXT:   (S) 0 num-cse'd - Number of operations CSE'd
 // PASSES-NEXT:   (S) 0 num-dce'd - Number of operations DCE'd
